Deploy_Streamlit/requirements.txt
pandas==2.1.1
numpy==1.26.0
google-auth==2.3.3
folium==0.14.0

